requirements.txt
flask
pinecone-client
openai
twilio
python-dotenv
requests
numpy
ngrok